Reformat all Java files with Google Formatter

+++ onebusaway-gtfs-transformer/src/test/java/org/onebusaway/gtfs_transformer/impl/AddExtensionFileTest.java
@@ -1,20 +1,26 @@
 /**
  * Copyright (C) 2023 Cambridge Systematics, Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * <p>Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License. You may obtain a copy of the License at
  *
- *         http://www.apache.org/licenses/LICENSE-2.0
+ * <p>http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
+ * <p>Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
 package org.onebusaway.gtfs_transformer.impl;
 
+import static org.junit.jupiter.api.Assertions.*;
+
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -24,18 +30,7 @@ import org.onebusaway.gtfs.serialization.GtfsWriter;
 import org.onebusaway.gtfs_transformer.AbstractTestSupport;
 import org.onebusaway.gtfs_transformer.services.TransformContext;
 
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-
-import static  org.junit.jupiter.api.Assertions.*;
-
-/**
- * test we can insert an extension into a GTFS file via a transformation.
- */
+/** test we can insert an extension into a GTFS file via a transformation. */
 public class AddExtensionFileTest extends AbstractTestSupport {
 
   private static final String TXT_STRING = "route_id,stop_id,direction_id,name\n";
@@ -52,8 +47,8 @@ public class AddExtensionFileTest extends AbstractTestSupport {
     _gtfs.putRoutes(1);
     _gtfs.putTrips(1, "r0", "sid0");
     _gtfs.putStopTimes("t0", "s0");
-
   }
+
   @AfterEach
   public void teardown() {
     _support.cleanup();
@@ -62,10 +57,10 @@ public class AddExtensionFileTest extends AbstractTestSupport {
   @Test
   public void run() throws IOException {
     File extensionFile = File.createTempFile("subwayRouteStop-", ".csv");
-    if (extensionFile.exists())
-      extensionFile.deleteOnExit();
+    if (extensionFile.exists()) extensionFile.deleteOnExit();
 
-    String extensionFilename = extensionFile.getAbsolutePath();;
+    String extensionFilename = extensionFile.getAbsolutePath();
+    ;
     BufferedWriter bwriter = new BufferedWriter(new FileWriter(extensionFilename));
     bwriter.write(TXT_STRING);
     bwriter.close();
@@ -77,8 +72,7 @@ public class AddExtensionFileTest extends AbstractTestSupport {
     test.run(context, dao);
 
     File tmpFileDirectory = File.createTempFile("AddExtensionFileTest-", "-tmp");
-    if (tmpFileDirectory.exists())
-      _support.deleteFileRecursively(tmpFileDirectory);
+    if (tmpFileDirectory.exists()) _support.deleteFileRecursively(tmpFileDirectory);
     tmpFileDirectory.mkdirs();
     _support.markForDeletion(tmpFileDirectory);
     writer.setOutputLocation(tmpFileDirectory);
@@ -92,6 +86,5 @@ public class AddExtensionFileTest extends AbstractTestSupport {
     assertTrue(expectedFile.isFile(), "expected extension to be a file!");
     String actualText = Files.readString(Path.of(modLocation));
     assertEquals(TXT_STRING, actualText);
-
   }
 }
